v3: Move the library, renaming it to acconeer.exptool
This only moves the files and thus breaks the project. Corresponding
changes to setup.py and all files referencing to acconeer_utils will
come in later commits.

The root dir is renamed from lib to src to be more in line with current
best practices.

For reference, see the PyPA sample Python project
https://github.com/pypa/sampleproject

Change-Id: I3383405cb66dbe60fd4c0df719a2081d0e4a260d