v3: Update setup.py and add MANIFEST.in
This also removes the previously added cleaning. The reason for that is
that it breaks the setup when using pip to install. The new recommended
way to do a clean install is to use a tool such as setupext-janitor. It
can (currently) be installed with pip. Using it, a clean install can be
done with:

python setup.py clean --all install --user

For reference, see the PyPA sample project:
https://github.com/pypa/sampleproject
and the Python packaging guidelines:
https://packaging.python.org

Change-Id: Iec2f957183edf79f2b9b289156aebca8465e20bc