v3: Replace start/stop_streaming with start/stop_session
The term streaming was introduced because of the streaming server (JSON
based over socket), but has since become legacy due to the addition of
the module server (register based over UART/SPI). The SPI (sub)protocol
does not support any form of streaming. Instead, the server is contin-
uously polled for data. Therefore it is misleading to use the term
streaming in general for the clients. As such, start/stop_streaming is
simply replaced with start/stop_session.

Change-Id: I98674f0552737789f2f402239423e12c1c675cd0