Revise the max leakage amplitude for parking
The default value of the parking detection parmeter leak_max_amplitude
is changed from 1000 to 2000 to handle the larger leakage that was
recorded for the temperature chamber setup with XC111-1 at -40°C for
other sensors than used in eariler tests.

A new paragraph is added in the documentation to describe this change.

The sensor that prompted this change is L019D018 (TT), and the
previously tested sensors are L019D019 (TT), L023D002 (FF),
L023D016 (FS), L019D005 (SF) and L023D013 (SS).

Feature: RS-3787
Change-Id: I533362b323a5d7212141a76abef554245e94ac5a