Change the way the update rate displayed in exptool is calculated
The update rate in the exptool shows the rate at
which the data is processed on the host.
This update changes the way the rate is calculated
to take an average of the last 8 values and then low-pass
filter it to avoid flickering numbers.
This will make it indicative of the rate at
which the exploration tool processes frames.

Change-Id: Ia6fd70c73cd69c9d6a760a0fa5369711fe54880b