Changes the detection threshold for the button press processing example
The detection threshold is quite low, which causes a noisy sensor to
produce false detections on the default settings.

Changing the threshold solves the issue, but makes detection very difficult
(again on the default settings) on the XM112.

The threshold is changed from 0.04 to 0.2. Only tested on XM132 and XM112.

Change-Id: Iafe3d1a119ef7e425fe6e0a3f974f6536b6262a6