Fixes a timing bug in tests.
The gui tests causes a crash for processors with low update rate.
A fix is to increase the wait time between tests,
so that the processor can deal with the returned data.

The increased timing requires the timeout limit to be increased as well.

Verified on separate commit as well.

Change-Id: I613ec5756165a9e51a8e5d634defa69dd9725972