Bugfix: A121: Distance: has_neighbour is false even if neighbour exist.
First element has_neighbour is False for second transition subsweep
as the conditon only checks whether or not a close range measurement
is included. This is a valid check for the first transition subsweep,
but not for the second.

Fix is to check if the length of the list of groups is different from zero.
If that is the case, then there exist a neighbour to the left of the segment.

Change-Id: I4a37839fc54e98009bee7a5e700a8b72c9036e3d