Add last error log from exploration server to error
The error revceived from the setup command typically does not say
that much about the problem (e.g. "RSS, could not get buffer size
(invalid config)") and instead the real error is sent as a log line
(e.g. "Start point or range too far").

This change tries to find this last error and if found add it to the
error message which will make it easier for the user to understand
the problem.

Change-Id: If6fbccfe2827bdf3c2c331eb12aa20d3ba195ab9