Bugfix: A121: Distance: Set context to None when calibrating.
Background: The shared state context is currently used when creating
the distance detector object for calibration. This becomes an issue
when switching from one sensor to another as the sensor id in the
supplied context and the sensor id from the GUI does not match.

Fix: Set context to None when calibrating. When the context supplied
to the distance detector is None, it will create a new context structure,
which is populated throughout the detector calibration process.

Change-Id: Id8184f23c7de3709ffeac4b1a5b05572d715cf9a