A121: Redesign client instantiation
Instead of wrapping exploration and mock clients in the Client class,
the Client.open() function is now a factory for the Client class
Depending on the ClientInfo contents an ExplorationClient or
a MockClient will be created.

This will change the public API for the Client creation from
client = Client() to client = Client.open().

Feature: RS-7238
Change-Id: I2f648ff62c8ddbd29d541848abead1327171cafb