Mypy: set disallow-any-generics
This flag disallows implicit Any-generics.
The reason we want to disallow this is because it's easy to forget the
type argument (GenericClass[TypeArguemnt]) and variables that have
type "Any" are invasive, as expressions
containing them will automatically be of type "Any"(!).

It's still allowed to explicitly set "Any" as the type parameter,
the flag & checks are there to make sure they are visible.

Change-Id: I6d75ad4fc3b3b615f1484098585f65bb1179ad58