Add tests for Backend
The Backend is the part of the application that handles
"tasks" (special tuples) to connect/disconnect a client and to
load/unload plugins. It runs in a separate process from the GUI.

Backend is the inner-most part of the application, followed by AppModel.

These tests form a specification of how the Backend responds
to the built-in tasks in different states.

Change-Id: I9a4ccfd1ca2d3d802dd72a327a3cc67c7ca7b354