Docs: Replace "working with the gui" section
Instead of the thorough guide of the Application, a separate section
for running ET and running examples was added.

Documentation like "working with the gui" (with pictures of the GUI)
was agreed to be hard to maintain (since the looks changes rapidly)
and not being very useful.

The A111 AlgorithmModule writer guide was moved
under the A111 Algorithms section.

Feature: RS-7475
Change-Id: I125ea3271de6f58a8c4d05e29c2b5de648556703