Fix bug where "on-set" errors did not display
The root cause was that validation errors cleared "on-set" errors
immediately.

This was mitigated by keeping track of erroneous aspects (members of
configs where the most recent set resulted in an "on-set" error)
until the same aspect is updated with a value that did not result in an
"on-set" error.

Also, "on-set" errors will always have precedence over validation errors.

Feature: RS-7294
Change-Id: Iaf5d3be55994f5d7b24dc7dce7504b733aa4e336