Make pidget's "hooks" parameter more ergonomic.
Technically, the "hooks" need to be stored in a tuple for PidgetGroups.
This requirement was earlier passed on to the user of the classes.

With this change, the user of the classes is free to pass a single
hook without a surrounding collection (seems common) or to pass multiple
hooks in a list.

Change-Id: I58ca739e14dd05c5df64824ea12a622d23ffc9b4