Begin replacing flake8 with ruff
This resulted of a speedup of "nox -s lint" by about 4x locally.

Ruff is a new linter that is super fast, which will shorten
feedback loops. It also offers a "--fix" option which makes "nox -s reformat":ing more potent.

It does not yet implement "flake8-future-annotations",
so that is kept for now.

Errors from "flake8-mutable" has never been shown(!) and the project
only cover built-in immutable types. The project haven't been updated
in the last 5 years, so it was scrapped.

Change-Id: Ib11204075486acb5c9861a917df4b28197422c22