A121: Distance: Bugfix: Incorrect end point with leakage cancellation
This bug was introduced with cba7636.

Prior to cba7636, a short end point resulted in the close range leakage
cancellation logic to be enforced.

With cba7636 implemented, it is possible to turn off the close range
cancellation logic, but not setting a short starting point. This change
fixes that.

Update test input/output-files as new logic result in different session
config.

Change-Id: I9003fd0179c6734624d329c03117903724a2357f