A121: Distance: Bugfix: Fix close range cancellation end point
When running in close range leakage cancellation mode, the end points
is always set to the transition point. This is incorrect. If the used
inputs a end point that is lower than the transition point, the
user provided input should be used.

Updated test input/output as this change result in a different session
config.

Change-Id: I78cdf0311f8a77b245726f0ae34b6b595e9ca6c2