A121: Update Client & Recorder API's
This change further separates the life-cycles of Client & Recorder.

Before, a Recorder recorded a single Client session.
The recording was initialized by passing a Recorder in
"Client.start_session" after which the Recorder was "stopped"
(file closed) in "Client.stop_session"

Now Recorder's life-cycle needs to be managed outside of the Client
(by using its context manager or ".close()"), which allows a Recorder
to record multiple client sessions. The recorder can still be attached
(just like the old "Client.start_session")
to a Client with "Client.attach_recorder".

Feature: RS-7594
Change-Id: I2cf841e0e853827059ff9b54f1892f26ba80e5b9