Update "is_task" decorator
"is_task" now wraps the passed method in a descriptor.
Aside from normal usage, the descriptor also provides a "rpc" method of
the descriptor. The kwargs of "rpc" is directly derived from the
wrapped method and is type checked as such.

All rpc implementations are identical, as they simply forward the
task name and arguments to be executed Backend-side.

rpcs are used to show previously hidden dependencies. The descriptor
is used in favour of regular decorators because of typing issues.

Feature: RS-6352
Change-Id: I36023713ea2c232f1212655432cef5c8b1d6258d