Add "extra_presenter" argument to AttrsConfigEditor
"extra_presenter" is a hook for different presentations in the
SaveDialog window, accessible via editors' save buttons (recently added).

A PresenterFunc accepts an instance and a PresentationType
(JSON or C_SET_CONFIG at the moment) and returns a non-None value if the
function can create a presentation matching the arguments.

This addition makes the presentations in the
SaveDialog window extendable from plugins.

Change-Id: I6901fa07c686568ef49b6b854ad0e9f3171d2300