Add placeholder text to OptionalFloatPidgetFactory
This is a configuration parameter for the OptionalFloatPidget that
enables a custom placeholder text when the data is None
(checkbox unchecked).

This config parameter can remove some ambiguity if used.
E.g. The placeholder text of SessionConfig.update_rate,
SensorConfig.sweep_rate & -.frame_rate was updated to say "- Hz" instead
of the grayed out arbitrary rates of 10Hz, 1000Hz & 10Hz, respectively.

Change-Id: I97440297f3d7e4022f720c2bc7fadb51be241d07