Refactor VersionLabel
VersionLabel was a clickable wrapper of a label and (optionally) an icon
with an overridden paintEvent, which made the label and the icon appear
as a single widget.

It was simplified by replacing the clickable wrapper by a regular button,
which supports both labels and icons natively.

The changelog popup was changed to be a dialog widget instead of a
regular widget.

Change-Id: Ib69b7300b38ca4185011aa5f3eae7f89192e7bef