Fallback to default processor config
This is done in absence of an algo group in a loaded file.

This interacts with the "fallback-load" of sparse_iq if a record
does not have an algo group in a way that records without an algo group
are played back in the GUI with sparse_iq with default config.

Change-Id: I5b2fc0c04497d3ee07e836a63cce40b1253d7376