Rework status message system
Instead of immediately overwriting the status like before, they are now
displayed one by one.

The message queue is run length encoded before displaying.

The display duration of a message is a sigmoid function of how many
status messages there in the queue. The duration is close to 0 when there
are 10 or more run length encoded messages in the queue.

Change-Id: I4c15dc675802bee4cddbc3ef46b03d339a6a98c4