PlotPlugin spring cleaning
Since there is nothing inherently A121-specific (or any other generation)
when plotting, the plot plugin baseclass was stripped of the A121 prefix,
and moved to the apps plugin base.

The same reasoning applies to the detector shim classes. The only thing
they do is adding indirection, which we can do without

Change-Id: I4896bd60373a6144cbf175f80fc0ab6b4b963919