Enable doctest
doctest is a standard module that runs simple tests
(disguised as docstring examples). The syntax is
close to what one would see in the python REPL.

This should make it more straightforward to add
simple tests and could also improve our documentation
if used (everyone loves examples).

This change includes a few examples of usage.

Change-Id: I533a1163a4ec3221037b2334ee5a09ed4c4ebd6f