Update PGProcess
This update gives the user a choice if it should setup
in the "layout" (old) style or the "widget" style.
If "layout" is used, the GraphicsLayout will be passed to
PGUpdater.setup and if "widget" is used, the GraphicsLayoutWidget
will be passed.

The flag "allow_subsampling" controls if the PGProcess is allowed
to skip results or not, but may cause applications to accumulate
lag, which an informative warning is added for.

Change-Id: Ic57d0c19268cefb15dd58c6f60a3bc55f8fa6639