Updates get_temperature_adjustment_factors
Adds more informative docstring and aligns output formats.
Also updates the input format to take two temperatures and calculate difference from them instead of allowing the user to do so.

The output format is aligned so the amplitude/noise deviation both is multiplied with the factor. Previously, the amplitude had to be divided.

Change-Id: I68520564808811837bab9023a4d78d507a81a08b