opser: Fix incompatibility with pickle
It seems MissingH5ObjectError being an Exception subclass
(which accepts a single str message to its constructor) played
poorly with pickle (ForkingPickler mentioned in the stacktrace).

This error must've been triggered when the MissingH5ObjectError
gets sent from the Backend and the main gui thread (where
objects get pickled as a part of inter-process communication).

All of this is circumventeted by using a factory method and
leaving the constructor as defined in Exception.

This also means that the Exception occuring in parking before
now gets surfaced.

Change-Id: I8fe6cddc012e7dae0fe26a3e563227786adaa25b