Waste level: Add validation check
The check that was added is that the subsweeps should
be monotonically increasing in start point.
The algorithm works without this check, so this is more
self-imposed than anything.

This also simplifies the overlap checks that have been
done. Now they can be checked linearly instead of checking
all combinations.

Change-Id: I88c7b21831701288d940c9dfd07a0dc758ae4beb