Replace doit and bash scripts with pytest
Recently found the "worker_id" fixture that enables
the calculation of unique ports per pytest-xdist test
worker.

This in conjunction with spawning exploration server
via pytest fixtures removes the need for both
the run-xxx-integration-tests.sh scripts and doit.

This does however remove the capability of running tests
on different python versions in parallel. Instead, it
allows the whole test suite to be ran with a single pytest
call, which could be interesting if restructuring the
Jenkins pipeline.

Change-Id: I5ead6275f7e80e1150f0249f73993d89ff105036