Add signal log to AppModelViewer
This is an extension to AppModelViewer that captures all
emits of the majority of AppModel's signals and puts a
string representation of the emitted object in a foldable view.

This should make it easier to debug race conditions as
all emits are shown. This is in contrast to only showing the
latest one, which was done before.

Change-Id: I8cac494c16c9da6879492478c02bfb95723435b6