ruff: Make sure that broad excepts are handled
This ruff rule checks for the pattern try-except-(pass|continue) and
prompts us to do something instead of nothing.

Some of the cases are exceptions that handle control flow, where an
ignore have been added, but the majority of places where a good place to
add a log (logging also had some rules added to it).

Change-Id: I98b9e99be0733892b140bded2b6656a90a410cb4