Error if no version file is present
This is a stricter handling of the generated version.
Previously, the fallback has only been observed triggering in Jenkins
(Not locally reproducible). Now an extra call to hatch has been added in
order to ensure the presence of '_version.py'

My hypothesis is that the version-fallback was caused by a
race-condition of 'git clean' and importing 'acconeer.exptool'
('_version.py' is not tracked in git).

The fallback ('0.0.0') has been fine to use until some time ago, where
we introduced version-dependant handling of recorded data (whether to do
rxmajor-to-txmajor translation). If the fallback is used, the
translation will be performed when it shouldn't.

Change-Id: Idd62d1901cda5c51704375d8f31b74ac20172069