Docs: Fix issue with autogenerated script pages
Each generated script page contains a link "View on GitHub".

If an example script is added, a generated script page is also added and
the Sphinx extension that finds broken links (linkcheck) correctly flags
the newly added GitHub link as broken.

This change makes the linkcheck ignore errors originating from the
generated script pages as the issue should resolve itself once the repo
is pushed to GitHub (and then RTD).

An alternative route is to remove the "View on GitHub" link, but I think
it's a nice link to have.

Change-Id: I89e460e164c0bda2791cb4dd27530bed007e5edd