type_migration: Correct TypeVar
'_MigT' describes the type(s) that can be migrated to the '_HeadT'.
'_MigT' accumulates types into a Union through 'epoch()' and 'load()'.

Note that 'Union[A] (= A)' is a subtype of 'Union[B, A]'.

E.g.:
  'str' is a subtype of 'Union[str, int]', implying that
  one may pass a 'str' into a function that expects a 'Union[str, int]'
  since 'str' is a subtype (Liskov Substitution Principle).

Making '_MigT' contravariant implies that
Epoch[X, Y, B] is a subtype of Epoch[X, Y, A]
given that A is a subtype of B (notice the flipped order).

This means that e.g.:
  'Epoch[X, Y, Union[str, int]]' is a subtype of 'Epoch[X, Y, str]' and
  one may pass a 'Epoch[X, Y, Union[str, int]]' into a function that
  expects 'Epoch[X, Y, str]' since 'Epoch[X, Y, Union[str, int]]' is a
  subtype (LSP).

Or in other words:
  It's ok to pass an 'Epoch[X, Y, Union[str, int]]' to a function
  expecting 'Epoch[X, Y, str]' since 'Epoch[X, Y, Union[str, int]]' is
  more "capable" than what's needed.

Change-Id: Ia9a84e3d51d7889a739c3f30b2967b275c67c5c5