distance: Small fix for loading cache in distance
Cache loading did not use type_migration, which would make us error in
more places than needed and the context timeline is already defined.

Also, If one were to only have a config in the cache (no context), the
migration will fail with a lengthy opser-error. This change catches that
case as well.

Change-Id: I6bd3fa857448caf0b66340302c1c475124d675f7