type_migration: Use ExceptionGroups
In order to gain insights into failures when using the type_migration
module, one needs to see all that went wrong during migration.

Before, only an opaque error of ~"Migration failed" was printed.

With this change, in case of a failure, a tree of all errors that
happened during will be displayed instead (this is where ExceptionGroups
come in).

Change-Id: I8fe3f9f3febb37437c26984964139111f815aab5