opser: Improve 'is_subclass'
This change tries to make 'is_subclass' more robust.

A version-dependant edge case was found which made this function
wrongfully throw an exception when called with the arguments

    is_subclass(list[SubsweepConfig], <SomeAlgoConfig>)

In order to handle this edge-case, it must be enforced that the
'superclass' argument is not a generic class (to not have to handle
covariant & contravariant type variables, which cannot be done
generally).

This constraint is already expressed in the type signature, but now also
in code for us to have control if something similar happens.

Change-Id: I64b87cf73b9b02d931a7276c48a7c3f783a7ddfb