Attempt to de-flake backend tests
This change aims to get rid of "false timeouts".

The backend tests are not ideal from the start as a separate Backend
process is spun up in order to do the tests.

'assert_messages' parameters for "total timeout", "timeout per receive"
and limit of received messages sole goal was to protect the tests suite
from hanging.

pytest-timeout offers a simpler way to guard against hung tests with a
simple module-wide timeout and reduces the risk of any of the
"smaller timeouts" being too short in certain circumstances.

Change-Id: I4c1bc60220fbd05b29bc9814d6b1b328ceb09e20