Bump docker to Ubuntu 24.04
An important difference is that the system python environment is more
protected. This means that e.g.

    pip3 install requests

is not the way to install the 'requests' package, but rather it should
be installed with something like

    sudo apt install python3-requests

I decided to utilize 'pipx' to manage the needed applications from
requirements-dev.txt (hatch & twine).
'pip' should not be needed (as things are handled through hatch
nowadays).
I can't see where 'requests' is used from, but if it is used, it should
be used from a hatch environment.

Change-Id: I3e2ac349e4ca43088df1bc37d35e67b8a867669d