Change version constraints of PySide6
PySide6==6.8.* had a rocky start, but also added support for Python 3.13

There are some prints that aren't explained and seems to come from deep
down in the stack. Let's hope it gets figured out and the prints
disappear in the future.

Apart from that, the application seems to work just fine.

Change-Id: Ife81ece250f3ebb96c3d0d68e8c42f9e50f4491d