opser: Make sure to validate Persistor names
The Persisor names is the location where a Group or Dataset is created.

In h5py, it's not allowed to create name-less Datasets, which is what
happends if one runs e.g.

    h5py.File(...).create_dataset("./", [5.0])

An error like this escaped in the case of serializing a bare 'None'
because the validation was simply not used.

Change-Id: Ic3f4a06873756497eb8598f02b12ff0e26bd0ec3