A121 update ET reg tests for level tracking.
Because the updates in tank level to enable partial window tracking
the assumption that 1 call to get_next() equals one frame in the
record does not hold any more. This is because the distance detector is
reinitialized every sample which triggers calibration frames breaking
this 1:1 assumption in the record. This made some of the regression
test to fail no matter the record structure. To fix this this assumpion
had to be removed in the test script. The strategy is now instead to
just call get_next() until the record is exhausted instead of a fixed
number of iterations.

Feature: RS-10682
Change-Id: Ie6508dc94207dff179e73cf62c853639a7967fe2