Rewrite Flash tab
This rewritten flash tab leverages the QWizard functionality of Qt,
which implements a classic setup wizard.

This is aimed to bring some clarity and order to the flashing process.

It's a bit yanky to press the big button in order to start the flashing
process, but that will be resolved in the upcoming changes.

Change-Id: Icd257c7810f6bafdadcd2d28a25b1da96fd20003