A121 backend plugin reg test update
When the tank level update rate is added the problem is that the log
for tank level now contains one session with one frame and that the
same session is lopped since the ref app calls start_session() each
get_next(). This means that the log is just looped until timeout. To
fix this its ensured that at least two sessions are logged--> session
index is not cycled. It was also realized that the underlying problem
is that we are handling older logs wich did only contain one session.
Therfore a second fix looking at the verions of the record was added
to prevent the cycling of the session index while replaying.

Feature: RS-10682
Change-Id: I6ce265c540a1e7f2b032dbd0aee265d6b1102544