opser: Ensure only Load-/SaveError is thrown from load/save
This change makes opser exhaust its current type tree (which is good)
but it is also more prone to swallow errors that might be interesting
(which is bad).

Said errors will still be part of the ExceptionGroup tree when opser
fails to serialize an instance (which is good).

The implementation changes 'load' and 'save' into template methods,
which should never be called directly.

Change-Id: I6e4181953c7b5f76fb86d9831d9ea531cce56b47